refactoring some methods
