+++ src/main/java/com/oceanprotocol/squid/manager/AgreementsManager.java
@@ -3,19 +3,32 @@ package com.oceanprotocol.squid.manager;
 import com.oceanprotocol.common.helpers.EncodingHelper;
 import com.oceanprotocol.common.web3.KeeperService;
 import com.oceanprotocol.squid.exceptions.ConditionNotFoundException;
+import com.oceanprotocol.squid.exceptions.EthereumException;
+import com.oceanprotocol.squid.exceptions.ServiceException;
 import com.oceanprotocol.squid.external.AquariusService;
 import com.oceanprotocol.squid.models.DDO;
+import com.oceanprotocol.squid.models.DID;
 import com.oceanprotocol.squid.models.service.Agreement;
 import com.oceanprotocol.squid.models.service.AgreementStatus;
 import com.oceanprotocol.squid.models.service.Service;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.web3j.abi.EventEncoder;
+import org.web3j.abi.datatypes.Event;
 import org.web3j.crypto.Keys;
+import org.web3j.protocol.core.DefaultBlockParameterName;
+import org.web3j.protocol.core.methods.request.EthFilter;
+import org.web3j.protocol.core.methods.response.EthLog;
 import org.web3j.protocol.core.methods.response.TransactionReceipt;
 import org.web3j.tuples.generated.Tuple7;
+import org.web3j.utils.Numeric;
 
+import java.io.IOException;
 import java.math.BigInteger;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class AgreementsManager extends BaseManager {
 
@@ -142,5 +155,100 @@ public class AgreementsManager extends BaseManager {
         else log.error("The current address" + address + "is not a condition address.");
         throw new ConditionNotFoundException("The current address" + address + "is not a condition address.");
     }
+
+
+
+    private List<DID> getAccessAgreementsFulfilledByConsumer(String consumerAddress) throws ServiceException {
+
+        EthFilter didFilter = new EthFilter(
+                DefaultBlockParameterName.EARLIEST,
+                DefaultBlockParameterName.LATEST,
+                accessSecretStoreCondition.getContractAddress()
+        );
+        try {
+
+            final Event event = accessSecretStoreCondition.FULFILLED_EVENT;
+            final String eventSignature = EventEncoder.encode(event);
+            didFilter.addSingleTopic(eventSignature);
+            didFilter.addNullTopic();
+            didFilter.addNullTopic();
+            didFilter.addOptionalTopics(Numeric.toHexStringWithPrefixZeroPadded(Numeric.toBigInt(consumerAddress), 64));
+
+            EthLog ethLog;
+
+            try {
+                ethLog = getKeeperService().getWeb3().ethGetLogs(didFilter).send();
+            } catch (IOException e) {
+                throw new EthereumException("Error creating consumedAssets filter.");
+            }
+
+            List<EthLog.LogResult> logs = ethLog.getLogs();
+            List<DID> DIDlist = new ArrayList<>();
+            for (int i = 0; i <= logs.size() - 1; i++) {
+                DIDlist.add(DID.getFromHash(Numeric.cleanHexPrefix((((EthLog.LogObject) logs.get(i)).getTopics().get(2)))));
+            }
+            return DIDlist;
+
+        } catch (Exception ex) {
+            log.error("\"Unable to retrieve access agreements fulfilled by " + consumerAddress + ex.getMessage());
+            throw new ServiceException("Unable to retrieve access agreements fulfilled by " + consumerAddress + ex.getMessage());
+        }
+
+    }
+
+
+    private List<DID> getComputeAgreementsFulfilledByConsumer(String consumerAddress) throws ServiceException {
+
+        EthFilter didFilter = new EthFilter(
+                DefaultBlockParameterName.EARLIEST,
+                DefaultBlockParameterName.LATEST,
+                computeExecutionCondition.getContractAddress()
+        );
+        try {
+
+            final Event event = computeExecutionCondition.FULFILLED_EVENT;
+            final String eventSignature = EventEncoder.encode(event);
+            didFilter.addSingleTopic(eventSignature);
+            didFilter.addNullTopic();
+            didFilter.addNullTopic();
+            didFilter.addOptionalTopics(Numeric.toHexStringWithPrefixZeroPadded(Numeric.toBigInt(consumerAddress), 64));
+
+            EthLog ethLog;
+
+            try {
+                ethLog = getKeeperService().getWeb3().ethGetLogs(didFilter).send();
+            } catch (IOException e) {
+                throw new EthereumException("Error creating consumedAssets filter.");
+            }
+
+            List<EthLog.LogResult> logs = ethLog.getLogs();
+            List<DID> DIDlist = new ArrayList<>();
+            for (int i = 0; i <= logs.size() - 1; i++) {
+                DIDlist.add(DID.getFromHash(Numeric.cleanHexPrefix((((EthLog.LogObject) logs.get(i)).getTopics().get(2)))));
+            }
+            return DIDlist;
+
+        } catch (Exception ex) {
+            log.error("Unable to retrieve compute agreements fulfilled by " + consumerAddress + ex.getMessage());
+            throw new ServiceException("Unable to retrieve compute agreements fulfilled by " + consumerAddress + ex.getMessage());
+        }
+
+    }
+
+    /**
+     * List of Asset objects purchased by consumerAddress
+     *
+     * @param consumerAddress ethereum address of consumer
+     * @return list of dids
+     * @throws ServiceException ServiceException
+     */
+    public List<DID> getConsumerAssets(String consumerAddress) throws ServiceException {
+
+        return Stream.concat(getAccessAgreementsFulfilledByConsumer(consumerAddress).stream(),
+                             getComputeAgreementsFulfilledByConsumer(consumerAddress).stream())
+                .collect(Collectors.toList());
+    }
+
+
 
 }
